172912 - Validation folder renamed to Validations AB#172912
