+++ src/UKHO.S100PermitService.Common/Validations/IUserPermitValidator.cs
@@ -1,7 +1,7 @@
 ﻿using FluentValidation.Results;
 using UKHO.S100PermitService.Common.Models.UserPermitService;
 
-namespace UKHO.S100PermitService.Common.Validation
+namespace UKHO.S100PermitService.Common.Validations
 {
     public interface IUserPermitValidator
     {
